render_version/requirements.txt
# =============================================================================
# TUELINH REAL ESTATE - MINIMAL REQUIREMENTS (TESTED & WORKING)
# =============================================================================

# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Environment & Configuration
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Image Processing & Media
Pillow>=10.0.1
easy-thumbnails>=2.8.5

# Static Files & Asset Management
whitenoise==6.6.0

# Form Styling & UI Enhancement
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Rich Text & Content Editing
django-ckeditor==6.7.0

# Enhanced Form Widgets
django-select2==8.1.2

# Data Display & Visualization
django-tables2==2.7.0
django-filter==23.4

# User Interface Enhancements
django-taggit==4.0.0

# Location & Communication
django-countries==7.5.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Production Server
gunicorn==21.2.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing Tools (Basic)
pytest==7.4.3
pytest-django==4.7.0
